Use symlink to share API params between python scripts
