+++ .github/docker-performance-tests/alarms-poller/Dockerfile
@@ -4,12 +4,12 @@ COPY . .
 
 RUN pip install docker boto3
 
-CMD python ./poll-during-performance-tests.py --cpu-load-threshold $CPU_LOAD_THRESHOLD \
-                                              --total-memory-threshold $TOTAL_MEMORY_THRESHOLD \
-                                              --logs-namespace $LOGS_NAMESPACE \
+CMD python ./poll_during_performance_tests.py --logs-namespace $LOGS_NAMESPACE \
                                               --metrics-period $HOSTMETRICS_INTERVAL_SECS \
                                               --num-of-cpus $NUM_OF_CPUS \
                                               --app-process-command-line-dimension-value "$APP_PROCESS_COMMAND_LINE_DIMENSION_VALUE" \
                                               --target-sha $TARGET_SHA \
                                               --github-run-id $GITHUB_RUN_ID \
+                                              --cpu-load-threshold $CPU_LOAD_THRESHOLD \
+                                              --total-memory-threshold $TOTAL_MEMORY_THRESHOLD \
                                               --image-suffix $IMAGE_SUFFIX
