+++ .github/docker-performance-tests/alarms-poller/metric_data_params.py
@@ -1,63 +1,15 @@
 import argparse
-import json
-import logging
-from datetime import datetime, timedelta
-from statistics import mean
-
-import boto3
-
-logging.basicConfig(
-    format="%(asctime)-8s %(levelname)-8s %(message)s",
-    level=logging.INFO,
-    datefmt="%FT%TZ",
-)
-
-logger = logging.getLogger(__file__)
 
 # AWS Client API Constants
 
 COMMIT_SHA_DIMENSION_NAME = "commit_sha"
 GITHUB_RUN_ID_DIMENSION_NAME = "github_run_id"
 PROCESS_COMMAND_LINE_DIMENSION_NAME = "process.command_line"
+
 METRIC_DATA_STATISTIC = "Sum"
 
 
-def parse_args():
-    parser = argparse.ArgumentParser(
-        description="""
-        produce-performance-test-results.py produces overall results for the
-        performance tests that were just run as JSON output.
-        """
-    )
-
-    parser.add_argument(
-        "--cpu-load-threshold",
-        required=True,
-        type=int,
-        help="""
-        The threshold the CPU Load (as a percentage) must stay under to not
-        trigger the alarm.
-
-        Examples:
-
-            --cpu-load-threshold=75
-        """,
-    )
-
-    parser.add_argument(
-        "--total-memory-threshold",
-        required=True,
-        type=int,
-        help="""
-        The threshold the Total Memory (in bytes) must stay under to not trigger
-        the alarm.
-
-        Examples:
-
-            --total-memory-threshold=$(echo 1.5 \* 2^30 | bc)
-        """,
-    )
-
+def add_arguments(parser: argparse.ArgumentParser):
     parser.add_argument(
         "--logs-namespace",
         required=True,
@@ -131,9 +83,9 @@ def parse_args():
         "--github-run-id",
         required=True,
         help="""
-        The Id for the current GitHub workflow run. Used to query Cloudwatch by
-        metric dimension value so metrics returned correspond to the app that
-        was performance tested.
+        The Id for the current GitHub workflow run. Used as a dimension by which
+        metrics are queried in Cloudwatch by so that metrics returned correspond
+        to the correct run of the app that was performance tested.
 
         Examples:
 
@@ -141,32 +93,8 @@ def parse_args():
         """,
     )
 
-    parser.add_argument(
-        "--test-duration-minutes",
-        required=True,
-        type=int,
-        help="""
-        The duration of the performance test, which is used to determine the
-        start of metrics to include in the snapshots.
 
-        Examples:
-
-            --test-duration-minutes=$(echo 1.5 \* 2^30 | bc)
-        """,
-    )
-
-    return parser.parse_args()
-
-
-if __name__ == "__main__":
-    logger.debug("Starting script to get performance test results.")
-
-    args = parse_args()
-
-    start_time = (
-        datetime.utcnow() - timedelta(minutes=args.test_duration_minutes)
-    ).strftime("%FT%TZ")
-
+def get_metric_data_params(args):
     cpu_load_metric_data_queries = [
         {
             "Id": "cpu_time_raw",
@@ -186,7 +114,7 @@ if __name__ == "__main__":
                         {
                             "Name": GITHUB_RUN_ID_DIMENSION_NAME,
                             "Value": args.github_run_id,
-                        }
+                        },
                     ],
                 },
                 "Stat": METRIC_DATA_STATISTIC,
@@ -223,7 +151,7 @@ if __name__ == "__main__":
                         {
                             "Name": GITHUB_RUN_ID_DIMENSION_NAME,
                             "Value": args.github_run_id,
-                        }
+                        },
                     ],
                 },
                 "Stat": METRIC_DATA_STATISTIC,
@@ -250,7 +178,7 @@ if __name__ == "__main__":
                         {
                             "Name": GITHUB_RUN_ID_DIMENSION_NAME,
                             "Value": args.github_run_id,
-                        }
+                        },
                     ],
                 },
                 "Stat": METRIC_DATA_STATISTIC,
@@ -268,49 +196,4 @@ if __name__ == "__main__":
         },
     ]
 
-    aws_client = boto3.client("cloudwatch")
-
-    metric_data_results = aws_client.get_metric_data(
-        StartTime=start_time,
-        EndTime=datetime.utcnow(),
-        MetricDataQueries=cpu_load_metric_data_queries
-        + total_memory_metric_data_queries,
-    )["MetricDataResults"]
-
-    benchmarks_json = json.dumps(
-        {
-            "benchmarks": [
-                {
-                    "Name": "Soak Test Average CPU Load",
-                    "Value": mean(
-                        next(
-                            metric_data
-                            for metric_data in metric_data_results
-                            if metric_data["Id"] == "cpu_load_expr"
-                        )["Values"]
-                    ),
-                    "Unit": "Percent",
-                },
-                {
-                    "Name": "Soak Test Average Virtual Memory Used",
-                    "Value": mean(
-                        next(
-                            metric_data
-                            for metric_data in metric_data_results
-                            if metric_data["Id"] == "total_memory_expr"
-                        )["Values"]
-                    )
-                    / (2 ** 20),
-                    "Unit": "Megabytes",
-                },
-            ]
-        },
-        indent=4,
-    )
-
-    logger.info("Found these benchmarks: %s", benchmarks_json)
-
-    with open("output.json", "w") as file_context:
-        file_context.write(benchmarks_json)
-
-    logger.info("Done producing Performance Test results.")
+    return cpu_load_metric_data_queries, total_memory_metric_data_queries
