+++ .github/scripts/performance-tests/get-metric-data/metric_data_params.py
@@ -1,0 +1,1 @@
+../../../../.github/docker-performance-tests/alarms-poller/metric_data_params.py
