+++ .github/workflows/soak-testing.yml
@@ -54,7 +54,7 @@ jobs:
         with:
           ref: ${{ env.TARGET_SHA }}
 
-      # MARK: - Instrumentation-Type Specific Values
+    # MARK: - Instrumentation-Type Specific Values
 
       # NOTE: The configuration of `APP_PROCESS_COMMAND_LINE_DIMENSION_VALUE` is
       # repo dependent
@@ -94,7 +94,7 @@ jobs:
           repository: open-telemetry/opentelemetry-python-contrib
           path: ${{ env.APP_PATH }}/opentelemetry-python-contrib
 
-      # MARK: - GitHub Workflow Event Type Specific Values
+    # MARK: - GitHub Workflow Event Type Specific Values
 
       - name: Use INPUT as commit SHA
         if: ${{ github.event_name == 'workflow_dispatch' }}
@@ -108,7 +108,7 @@ jobs:
         run: |
           echo "TEST_DURATION_MINUTES=${{ github.event.inputs.test_duration_minutes || env.DEFAULT_TEST_DURATION_MINUTES }}" | tee --append $GITHUB_ENV;
 
-      # MARK: - Build and Push Sample App
+    # MARK: - Build and Push Sample App
 
       # WARNING: In an edge case, if two jobs use the same commit at the same
       # time, they will share an `IMAGE_SUFFIX` and race to push and pull Docker
@@ -156,7 +156,7 @@ jobs:
           cache-from: type=local,src=/tmp/.buildx-cache
           cache-to: type=local,dest=/tmp/.buildx-cache
 
-      # MARK: - Run Performance Tests
+    # MARK: - Run Performance Tests
 
       - name: Configure AWS Credentials
         uses: aws-actions/configure-aws-credentials@v1
@@ -192,24 +192,25 @@ jobs:
           # AWS_DEFAULT_REGION
           # IMAGE_SUFFIX
           # GITHUB_RUN_ID
-        run: >-
+        run: |-
           docker-compose up --build;
-          RUN_EXIT_CODE=$(docker inspect $(docker ps --quiet --all --filter "name=docker-performance-tests_alarms-poller") --format="{{.State.ExitCode}}");
-          echo "::set-output name=run_exit_code::$RUN_EXIT_CODE";
-          exit $RUN_EXIT_CODE;
+          echo "RUN_TESTS_EXIT_CODE=$(
+            docker inspect $(
+              docker ps --quiet --all --filter "name=docker-performance-tests_alarms-poller"
+            ) --format="{{.State.ExitCode}}"
+          )" | tee --append $GITHUB_ENV;
+          exit ${{ env.RUN_TESTS_EXIT_CODE }}
       - name: Fail early if Soak Tests failed to start
-        if: ${{ steps.check-failure-during-performance-tests.outputs.run_exit_code == '1' }}
+        if: ${{ env.RUN_TESTS_EXIT_CODE == 1 }}
         run: exit 1;
 
-      # MARK: - Report on Performance Test Results
+    # MARK: - Report on Performance Test Results
 
       - name: Install script dependencies
         run: pip install boto3
       - name: Get a snapshot of metrics and commit them to the repository
         run: |
-          python3 .github/scripts/performance-tests/produce-metric-widget-images.py \
-            --cpu-load-threshold ${{ env.CPU_LOAD_THRESHOLD }} \
-            --total-memory-threshold ${{ env.TOTAL_MEMORY_THRESHOLD }} \
+          python3 .github/scripts/performance-tests/produce_metric_widget_images.py \
             --logs-namespace ${{ env.LOGS_NAMESPACE }} \
             --metrics-period ${{ env.HOSTMETRICS_INTERVAL_SECS }} \
             --num-of-cpus ${{ env.NUM_OF_CPUS }} \
@@ -217,6 +218,8 @@ jobs:
             --target-sha ${{ env.TARGET_SHA }} \
             --github-run-id ${GITHUB_RUN_ID} \
             --test-duration-minutes ${{ env.TEST_DURATION_MINUTES }} \
+            --cpu-load-threshold ${{ env.CPU_LOAD_THRESHOLD }} \
+            --total-memory-threshold ${{ env.TOTAL_MEMORY_THRESHOLD }} \
             --app-platform ${{ matrix.app-platform }} \
             --instrumentation-type ${{ matrix.instrumentation-type }} \
             --max-benchmarks-to-keep ${{ env.MAX_BENCHMARKS_TO_KEEP }} \
@@ -230,16 +233,14 @@ jobs:
           git push;
           git checkout main;
       - name: Prepare Performance Test results as JSON output
-        run: python3 .github/scripts/performance-tests/produce-performance-test-results.py
-          --cpu-load-threshold ${{ env.CPU_LOAD_THRESHOLD }}
-          --total-memory-threshold ${{ env.TOTAL_MEMORY_THRESHOLD }}
+        run: python3 .github/scripts/performance-tests/get-metric-data/produce_performance_test_results.py
           --logs-namespace ${{ env.LOGS_NAMESPACE }}
           --metrics-period ${{ env.HOSTMETRICS_INTERVAL_SECS }}
           --num-of-cpus ${{ env.NUM_OF_CPUS }}
           --app-process-command-line-dimension-value "${{ env.APP_PROCESS_COMMAND_LINE_DIMENSION_VALUE }}"
-          --test-duration-minutes ${{ env.TEST_DURATION_MINUTES }}
           --target-sha ${{ env.TARGET_SHA }}
           --github-run-id ${GITHUB_RUN_ID}
+          --test-duration-minutes ${{ env.TEST_DURATION_MINUTES }}
       - name: Is this the first time we run Performance Tests for this commit?
         continue-on-error: true
         id: check-is-first-time-reporting-commit
